requirements.txt
osmium
numpy
matplotlib
scipy
alphashape
Pillow
pyyaml